Summary: fix for Bug 1924965- Build errors in Windows for 64 bit libraries
Reviewer: bravindra, done

Details: casting to int is required in the Find string template functions

git-svn-id: svn://svn.code.sf.net/p/framewave/code/trunk@208 f364fc6c-7e41-0410-b16c-eb8c68fe7df8

+++ Framewave/domain/fwSignal/src/Find.cpp
@@ -32,7 +32,7 @@ namespace OPT_LEVEL
 
                 if(strend == str)
                 {
-                    index = (src - lenfind) - start ;
+                    index = ((int)src - lenfind) - (int)start ;
                     return index;
                 }
 
@@ -61,7 +61,7 @@ namespace OPT_LEVEL
 
                 if(strend == str && *strend == *src)
                 {
-                    index = (end- start - lenfind + 1) ;
+                    index = ((int)(end- start) - lenfind + 1) ;
                     return index;
                 }
             }
@@ -83,7 +83,7 @@ namespace OPT_LEVEL
 
             if(substr == *srcstr)
             {
-                index = srcstr - start;
+                index = (int)(srcstr - start);
             }
             return index;
         }
@@ -100,7 +100,7 @@ namespace OPT_LEVEL
 
             if(substr == *end)
             {
-                index = end - start;
+                index = (int)(end - start);
             }
             return index;
         }
